requirementsDev.txt
ansible==5.2.0
molecule==3.5.2
molecule-docker==1.1.0
pytest==6.2.5
pytest-testinfra==6.5.0
